% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scroll.R
\docType{package}
\name{scroll}
\alias{scroll-package}
\alias{scroll}
\title{scroll}
\usage{
scroll(df)
}
\arguments{
\item{df}{A data frame.}
}
\description{
Scroll through a data frame's rows and columns using less..
}
\author{
\strong{Maintainer}: Pieter Provoost \email{pieterprovoost@gmail.com}

}
