% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffold.R
\name{scroll_config}
\alias{scroll_config}
\title{Read a scroll project config}
\usage{
scroll_config(dir)
}
\arguments{
\item{dir}{A scroll project directory.}
}
\value{
The parsed config as a list.
}
\description{
Read a scroll project config
}
