% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{scroll_manifest}
\alias{scroll_manifest}
\title{Read a scroll project manifest}
\usage{
scroll_manifest(dir)
}
\arguments{
\item{dir}{A scroll project directory.}
}
\value{
The parsed manifest as a list.
}
\description{
Read a scroll project manifest
}
