% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve.R
\name{scroll_render}
\alias{scroll_render}
\title{Render a scroll project for deployment}
\usage{
scroll_render(dir = ".", ...)
}
\arguments{
\item{dir}{A scroll project directory.}

\item{...}{Extra arguments passed to \verb{quarto render}.}
}
\value{
Invisibly, the exit status.
}
\description{
Render a scroll project for deployment
}
