% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{scroll_reset_panels}
\alias{scroll_reset_panels}
\title{Clear all custom panels registered with \code{\link[=register_panel]{register_panel()}}}
\usage{
scroll_reset_panels()
}
\value{
Invisibly, \code{NULL}.
}
\description{
Clear all custom panels registered with \code{\link[=register_panel]{register_panel()}}
}
