% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{scroll_view_kind}
\alias{scroll_view_kind}
\title{Which output kind a view produces}
\usage{
scroll_view_kind(view_type)
}
\arguments{
\item{view_type}{A view type name.}
}
\value{
\code{"plot"} or \code{"table"}.
}
\description{
Which output kind a view produces
}
