% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{scroll_disconnect}
\alias{scroll_disconnect}
\title{Close a scroll duckdb connection}
\usage{
scroll_disconnect(con)
}
\arguments{
\item{con}{A connection from \code{\link[=scroll_connect]{scroll_connect()}}.}
}
\description{
Close a scroll duckdb connection
}
