% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{view_umap_colorby}
\alias{view_umap_colorby}
\title{Embedding colored by a metadata column}
\usage{
view_umap_colorby(cells, params)
}
\arguments{
\item{cells}{The globally-loaded cells data.frame.}

\item{params}{List with \code{embedding} and \code{color_by}.}
}
\value{
A ggplot.
}
\description{
Embedding colored by a metadata column
}
